requirements.txt
amqp==2.5.2
billiard==3.6.1.0
celery==4.4.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.4
django-audiofield==0.10.0
django-celery==3.3.1
django-datetimepicker==3.14
django-extensions==2.2.5
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
gunicorn==20.0.4
importlib-metadata==1.3.0
kombu==4.6.7
more-itertools==8.0.2
mysql-connector-python==8.0.18
mysqlclient==1.4.4
Pillow==6.2.1
protobuf==3.10.0
PyMySQL==0.9.3
pytz==2019.3
six==1.13.0
static3==0.7.0
vine==1.3.0
xlrd==1.2.0
zipp==0.6.0